Making generation faster with better dropoff evaluation

+++ Assets/OpenSimplexNoise.cs
@@ -9,22 +9,22 @@ using System.Runtime.CompilerServices;
 
 public class OpenSimplexNoise
 {
-    private const double STRETCH_2D = -0.211324865405187;    //(1/Math.sqrt(2+1)-1)/2;
-    private const double STRETCH_3D = -1.0 / 6.0;            //(1/Math.sqrt(3+1)-1)/3;
-    private const double STRETCH_4D = -0.138196601125011;    //(1/Math.sqrt(4+1)-1)/4;
-    private const double SQUISH_2D = 0.366025403784439;      //(Math.sqrt(2+1)-1)/2;
-    private const double SQUISH_3D = 1.0 / 3.0;              //(Math.sqrt(3+1)-1)/3;
-    private const double SQUISH_4D = 0.309016994374947;      //(Math.sqrt(4+1)-1)/4;
-    private const double NORM_2D = 1.0 / 47.0;
-    private const double NORM_3D = 1.0 / 103.0;
-    private const double NORM_4D = 1.0 / 30.0;
+    private const float STRETCH_2D = -0.211324865405187f;    //(1/Math.sqrt(2+1)-1)/2;
+    private const float STRETCH_3D = -1.0f / 6.0f;            //(1/Math.sqrt(3+1)-1)/3;
+    private const float STRETCH_4D = -0.138196601125011f;    //(1/Math.sqrt(4+1)-1)/4;
+    private const float SQUISH_2D = 0.366025403784439f;      //(Math.sqrt(2+1)-1)/2;
+    private const float SQUISH_3D = 1.0f / 3.0f;              //(Math.sqrt(3+1)-1)/3;
+    private const float SQUISH_4D = 0.309016994374947f;      //(Math.sqrt(4+1)-1)/4;
+    private const float NORM_2D = 1.0f / 47.0f;
+    private const float NORM_3D = 1.0f / 103.0f;
+    private const float NORM_4D = 1.0f / 30.0f;
 
     private byte[] perm;
     private byte[] perm2D;
     private byte[] perm3D;
     private byte[] perm4D;
 
-    private static double[] gradients2D = new double[]
+    private static float[] gradients2D = new float[]
     {
              5,  2,    2,  5,
             -5,  2,   -2,  5,
@@ -32,7 +32,7 @@ public class OpenSimplexNoise
             -5, -2,   -2, -5,
     };
 
-    private static double[] gradients3D =
+    private static float[] gradients3D =
     {
             -11,  4,  4,     -4,  11,  4,    -4,  4,  11,
              11,  4,  4,      4,  11,  4,     4,  4,  11,
@@ -44,7 +44,7 @@ public class OpenSimplexNoise
              11, -4, -4,      4, -11, -4,     4, -4, -11,
         };
 
-    private static double[] gradients4D =
+    private static float[] gradients4D =
     {
              3,  1,  1,  1,      1,  3,  1,  1,      1,  1,  3,  1,      1,  1,  1,  3,
             -3,  1,  1,  1,     -1,  3,  1,  1,     -1,  1,  3,  1,     -1,  1,  1,  3,
@@ -183,7 +183,7 @@ public class OpenSimplexNoise
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private static int FastFloor(double x)
+    private static int FastFloor(float x)
     {
         var xi = (int)x;
         return x < xi ? xi - 1 : xi;
@@ -224,12 +224,12 @@ public class OpenSimplexNoise
         }
     }
 
-    public double Evaluate(double x, double y, int octaves, double persistence = .5f, double lacunarity = 2)
+    public float Evaluate(float x, float y, int octaves, float persistence = .5f, float lacunarity = 2)
     {
-        double total = 0;
-        double frequency = 1;
-        double amplitude = 1;
-        double totalAmplitude = 0;  // Used for normalizing result to 0.0 - 1.0
+        float total = 0;
+        float frequency = 1;
+        float amplitude = 1;
+        float totalAmplitude = 0;  // Used for normalizing result to 0.0 - 1.0
         for (int i = 0; i < octaves; i++)
         {
             total += Evaluate(x * frequency, y * frequency) * amplitude;
@@ -240,7 +240,7 @@ public class OpenSimplexNoise
         return total / totalAmplitude;
     }
 
-    public double Evaluate(double x, double y)
+    public float Evaluate(float x, float y)
     {
         var stretchOffset = (x + y) * STRETCH_2D;
         var xs = x + stretchOffset;
@@ -266,7 +266,7 @@ public class OpenSimplexNoise
 
         var c = lookup2D[hash];
 
-        var value = 0.0;
+        var value = 0.0f;
         while (c != null)
         {
             var dx = dx0 + c.dx;
@@ -288,7 +288,7 @@ public class OpenSimplexNoise
         return value * NORM_2D;
     }
 
-    public double Evaluate(double x, double y, double z)
+    public float Evaluate(float x, float y, float z)
     {
         var stretchOffset = (x + y + z) * STRETCH_3D;
         var xs = x + stretchOffset;
@@ -321,7 +321,7 @@ public class OpenSimplexNoise
 
         var c = lookup3D[hash];
 
-        var value = 0.0;
+        var value = 0.0f;
         while (c != null)
         {
             var dx = dx0 + c.dx;
@@ -346,7 +346,7 @@ public class OpenSimplexNoise
         return value * NORM_3D;
     }
 
-    public double Evaluate(double x, double y, double z, double w)
+    public float Evaluate(float x, float y, float z, float w)
     {
         var stretchOffset = (x + y + z + w) * STRETCH_4D;
         var xs = x + stretchOffset;
@@ -387,7 +387,7 @@ public class OpenSimplexNoise
 
         var c = lookup4D[hash];
 
-        var value = 0.0;
+        var value = 0.0f;
         while (c != null)
         {
             var dx = dx0 + c.dx;
@@ -416,11 +416,11 @@ public class OpenSimplexNoise
 
     private class Contribution2
     {
-        public double dx, dy;
+        public float dx, dy;
         public int xsb, ysb;
         public Contribution2 Next;
 
-        public Contribution2(double multiplier, int xsb, int ysb)
+        public Contribution2(float multiplier, int xsb, int ysb)
         {
             dx = -xsb - multiplier * SQUISH_2D;
             dy = -ysb - multiplier * SQUISH_2D;
@@ -431,11 +431,11 @@ public class OpenSimplexNoise
 
     private class Contribution3
     {
-        public double dx, dy, dz;
+        public float dx, dy, dz;
         public int xsb, ysb, zsb;
         public Contribution3 Next;
 
-        public Contribution3(double multiplier, int xsb, int ysb, int zsb)
+        public Contribution3(float multiplier, int xsb, int ysb, int zsb)
         {
             dx = -xsb - multiplier * SQUISH_3D;
             dy = -ysb - multiplier * SQUISH_3D;
@@ -448,11 +448,11 @@ public class OpenSimplexNoise
 
     private class Contribution4
     {
-        public double dx, dy, dz, dw;
+        public float dx, dy, dz, dw;
         public int xsb, ysb, zsb, wsb;
         public Contribution4 Next;
 
-        public Contribution4(double multiplier, int xsb, int ysb, int zsb, int wsb)
+        public Contribution4(float multiplier, int xsb, int ysb, int zsb, int wsb)
         {
             dx = -xsb - multiplier * SQUISH_4D;
             dy = -ysb - multiplier * SQUISH_4D;
